AGENT/requirements.txt
pydantic-ai
jinja2
httpx
python-dotenv
streamlit
watchdog
